refactor: make flat architecture for tag-cell-edit
