--- src/react/table-app/tag-cell-edit/component/MenuBody/styles.css
@@ -1,4 +1,0 @@
-.NLT__tag-menu-body {
-    max-height: 140px;
-    overflow-y: scroll;
-}
