--- src/react/table-app/tag-cell-edit/component/MenuHeader/styles.css
@@ -1,4 +1,0 @@
-.NLT__tag-menu-header {
-    background-color: rgba(0, 0, 0, 0.25);
-    padding: 4px 10px;
-}
