--- src/react/table-app/tag-cell-edit/component/SelectableTag/styles.css
@@ -1,7 +1,0 @@
-.NLT__selectable-tag {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: var(--nlt-spacing--sm) var(--nlt-spacing--md);
-    overflow: hidden;
-}
