+++ src/react/table-app/tag-cell-edit/selectable-tag.tsx
@@ -3,7 +3,6 @@ import TagColorMenu from "src/react/table-app/tag-color-menu";
 import { MenuLevel, MenuPosition } from "src/shared/menu/types";
 import { useMenu } from "src/shared/menu/hooks";
 
-import "./styles.css";
 import { Button } from "src/react/shared/button";
 import Icon from "src/react/shared/icon";
 import Tag from "src/react/shared/tag";
